requirements.txt
PySide6==6.10.1
requests==2.32.5
hjson==3.1.0
pyahocorasick==2.2.0